Endianness
Added endian order swapping. First completed the reverse bytes and used to create the swap order. Still wondering if the interface makes sense for swap. Also, not sure how to make a useful test for swap that isnt basically the same as reverse. Except that maybe it would confirm that the interface was present...